Using TS project references in integration tests
