+++ integration-tests/tests/package.json
@@ -2,6 +2,7 @@
   "name": "@realm/integration-tests",
   "version": "0.1.0",
   "description": "A set of tests that can run in different environments",
+  "type": "module",
   "main": "src/index.ts",
   "exports": {
     ".": "./src/index.ts",
@@ -24,7 +25,7 @@
       ]
     },
     "build": {
-      "command": "tsc",
+      "command": "tsc --build .",
       "dependencies": [
         "build-dependencies"
       ]
