+++ integration-tests/tests/src/node/inject-globals.ts
@@ -18,34 +18,35 @@
 
 // Set the defult depth of objects logged with console.log to improve DX when debugging
 import { inspect } from "node:util";
+import { existsSync } from "node:fs";
+import { dirname, resolve } from "node:path";
+
 inspect.defaultOptions.depth = null;
 
-if (!global.gc) {
+if (typeof gc !== "function") {
   throw new Error("Run with --expose_gc to allow garbage collection between tests");
 }
 
 // Require this file to get the Realm constructor injected into the global.
 // This is only useful when we want to run the tests outside of any particular environment
 
-import { existsSync } from "node:fs";
-global.fs = {
-  exists(path) {
-    return existsSync(path);
+Object.assign(global, {
+  title: "Realm JS development-mode",
+  environment: { node: true },
+  fs: {
+    exists(path: string) {
+      return existsSync(path);
+    },
   },
-};
-
-import { dirname, resolve } from "path";
-global.path = {
-  dirname(path) {
-    return dirname(path);
+  path: {
+    dirname(path: string) {
+      return dirname(path);
+    },
+    resolve(...paths: string[]) {
+      return resolve(...paths);
+    },
   },
-  resolve(...paths: string[]) {
-    return resolve(...paths);
-  },
-};
-
-global.title = "Realm JS development-mode";
-global.environment = { node: true };
+});
 
 function parseValue(value: string | undefined) {
   if (typeof value === "undefined" || value === "true") {
