--- integration-tests/tests/src/node/tsconfig.json
@@ -1,9 +1,0 @@
-{
-  "extends": "../../tsconfig.json",
-  "compilerOptions": {
-    "types": [
-      "node"
-    ]
-  },
-  "exclude": []
-}
